Separate web authentication responsibilities
