+++ .github/workflows/full-e2e.yml
@@ -50,7 +50,7 @@ jobs:
               - 'app/templates/**'
               - 'app/static/**'
               - 'app/modules/**'
-              - 'app/web.py'
+              - 'app/web*.py'
               - 'app/blueprints/**'
               - 'app/tz.py'
               - 'app/i18n/**'
