+++ .github/workflows/test.yml
@@ -68,7 +68,7 @@ jobs:
               - 'app/templates/**'
               - 'app/static/**'
               - 'app/modules/**'
-              - 'app/web.py'
+              - 'app/web*.py'
               - 'app/blueprints/**'
               - 'app/tz.py'
               - 'app/i18n/**'
