+++ app/modules/journal/routes.py
@@ -12,10 +12,10 @@ from flask import Blueprint, request, jsonify, send_file, make_response
 
 from app.tz import local_date_to_utc_range
 from app.web import (
-    require_auth,
     get_config_manager, get_state, get_storage,
-    _valid_date, _localize_timestamps, _get_client_ip, _get_lang, _get_tz_name,
+    _valid_date, _localize_timestamps, _get_lang, _get_tz_name,
 )
+from app.web_auth import require_auth, _get_client_ip
 from app.storage import ALLOWED_MIME_TYPES, MAX_ATTACHMENT_SIZE, MAX_ATTACHMENTS_PER_ENTRY
 
 from .import_parser import parse_file
